Rename and update map-layer docs
